requirements.txt
mlx
mlx-lm
numpy
tqdm
huggingface-hub
python-dotenv